backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
tensorflow>=2.15.0
Pillow>=10.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
